refactor: Move getting UTC time to a separate module
Replaces all use of `OffsetDateTime::now_utc()` with
`crate::time::now_utc()` which under the hood switched to using
`UtcDateTime`.

Mocking of UTC time is now packed away in the `time` module.
